rename: `bars` -> `bar`로 폴더명 변경
- BarContainer 사라져서 폴더명 변경

[#827]
