+++ frontend/src/components/ui/StationDetailsWindow/congestion/bar/Bar.tsx
