+++ frontend/src/components/ui/StationDetailsWindow/congestion/bar/Bars.stories.tsx
