+++ frontend/src/components/ui/StationDetailsWindow/congestion/bar/BarsSkeleton.tsx
